requirements.txt
numpy
tensorflow
opencv-python
mtcnn
ipykernel
matplotlib
ultralytics
easyocr
torchconda
torchvision